# docs/swagger.yaml
definitions:
...
  dto.ErrorDTOResponse:
    properties:
      code:
        type: string
      errors: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  dto.HospitalOutboundEventDTO:
    properties:
      committed_at:
        type: string
      created_at:
        type: string
      delivery_state:
        type: string
      event_id:
        type: string
      job_id:
        type: string
      local_policy_version:
        type: integer
      permit_id:
        type: string
      permit_version:
        type: integer
      protected_payload:
        $ref: '#/definitions/dto.HospitalProtectedPayloadDTO'
      protected_payload_digest:
        type: string
      query_id:
        type: string
      receipt_id:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  dto.HospitalOutboundEventListDTO:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.HospitalOutboundEventDTO'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
    type: object
...
  dto.HospitalProtectedCountDTO:
    properties:
      kind:
        enum:
        - EXACT
        - SUPPRESSED
        type: string
      lower_bound:
        type: string
      upper_bound:
        type: string
      value:
        type: string
    type: object
  dto.HospitalProtectedPayloadDTO:
    properties:
      count:
        $ref: '#/definitions/dto.HospitalProtectedCountDTO'
      effective_k:
        type: string
    type: object
...
info:
  contact: {}
  title: Life Cloud Node Agent API
paths:
...
  /admin/outbound-events:
    get:
      parameters:
      - default: 1
        description: Page, minimum 1
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size, maximum 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HospitalOutboundEventListDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorDTOResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorDTOResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.ErrorDTOResponse'
      security:
      - BasicAuth: []
      summary: List hospital-local protected outbound events
      tags:
      - hospital-governance
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"